+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## FUTURE
 
+* [CHANGED] - config file for `production` environment has been moved to 'config/production/' subdirectory
 * [REMOVED] - handling of shell process arguments (the arguments are processed directly by bi-service@1.0.0)
 
 ## 1.2.1